Rename functions folder to extra_functions
