+++ app/extra_functions/notification_funct.py
